requirements.txt
fastapi==0.129.0
uvicorn==0.41.0
starlette==0.52.1
pydantic==2.12.5
typing-extensions==4.15.0
typing-inspection==0.4.2
annotated-doc==0.0.4
click==8.3.1
h11==0.16.0
colorama==0.4.6
annotated-types==0.7.0
pydantic-core==2.41.5
anyio==4.12.1
idna==3.11
SQLAlchemy==2.0.46
greenlet==3.3.1
alembic==1.18.4
Mako==1.3.10
MarkupSafe==3.0.3
asyncpg==0.31.0
python-dotenv==1.2.1
psycopg2==2.9.11


